requirements.txt
cryptography
email-validator
sqlmodel~=0.0.24
psycopg2-binary
fastapi~=0.115.12
uvicorn~=0.34.2
passlib[bcrypt]~=1.7.4
pydantic~=2.11.4
SQLAlchemy~=2.0.40
python-dotenv~=1.1.0
python-jose[cryptography]~=3.4.0
python-multipart
PyJWT~=2.10.1